% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasetsExamples.R
\docType{data}
\name{SCE.colon.Lee}
\alias{SCE.colon.Lee}
\title{Colorectal cancer single-cell RNA-seq dataset from Lee, Hong,
Etlioglu Cho et al., 2020}
\format{
An object of class \code{SingleCellExperiment} with 2000 rows and 106364 columns.
}
\source{
\url{https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE75688}
\url{https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE132257}
\url{https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE144735}
}
\usage{
SCE.colon.Lee
}
\description{
\code{SingleCellExperiment} object containing single-cell RNA-seq data from
colorectal cancer samples from the Lee, Hong, Etlioglu Cho et al., 2020
study (GSE132465, GSE132257 and GSE144735).The raw count matrix only contains
information for the 2,000 genes selected to build the \code{DDLS.colon.lee}
pre-trained model from the \code{digitalDLSorteRmodels} R package.
}
\references{
Lee HO, Hong Y, Etlioglu HE, Cho YB, Pomella V, Van den Bosch B,
Vanhecke J, Verbandt S, Hong H, Min JW, Kim N, Eum HH, Qian J, Boeckx B,
Lambrechts D, Tsantoulis P, De Hertogh G, Chung W, Lee T, An M, Shin HT,
Joung JG, Jung MH, Ko G, Wirapati P, Kim SH, Kim HC, Yun SH, Tan IBH,
Ranjan B, Lee WY, Kim TY, Choi JK, Kim YJ, Prabhakar S, Tejpar S, Park WY.
Lineage-dependent gene expression programs influence the immune landscape
of colorectal cancer. Nat Genet. 2020 Jun;52(6):594-603.
doi: \doi{10.1038/s41588-020-0636-z}.
}
\keyword{datasets}
