% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasetsExamples.R
\docType{data}
\name{TCGA.colon.se}
\alias{TCGA.colon.se}
\title{Colorectal cancer bulk RNA-Seq samples from TCGA Research Network}
\format{
An object of class \code{SummarizedExperiment} with 59085 rows and 521 columns.
}
\source{
\url{https://www.cancer.gov/about-nci/organization/ccg/research/structural-genomics/tcga}
}
\usage{
TCGA.colon.se
}
\description{
Colorectal cancer bulk RNA-Seq samples from TCGA Research Network. These data
is provided as raw counts.
}
\references{
Koboldt, D. C., Fulton, R. S., McLellan, M. D., Schmidt, H.,
Kalicki-Veizer, J., McMichael, J. F., et al. (2012). Comprehensive
molecular portraits of human breast tumours. Nature 490 (7418), 61–70. doi:
10.1038/nature11412

Ciriello, G., Gatza, M. L., Beck, A. H., Wilkerson, M. D., Rhie, S. K.,
Pastore, A., et al. (2015). Comprehensive molecular portraits of invasive
lobular breast cancer. Cell. 163 (2), 506–519. doi:
10.1016/j.cell.2015.09.033

Torroja, C. and Sánchez-Cabo, F. (2019). digitalDLSorter: A Deep Learning
algorithm to quantify immune cell populations based on scRNA-Seq data.
Frontiers in Genetics 10, 978. doi: \doi{10.3389/fgene.2019.00978}
}
\keyword{datasets}
